Increment nucleic acid abbreviation counter for CMO samples
From investigator:

```
Counter: integer value incremented when the source is the same. (eg. if C-123456-T001-d is sequenced on Agilent Exome
and now the investigator wants to use the same sample with IDT, then we should have the sample labled C-123456-T001-
d02). Counter will be a 2 digit integer value range from 01-99 (values less < 10 are filled in with zeros '0' to preserve 2-digit
format). From the time of implementation the first sample for a particular Nucleic Acid get 01.
```

Signed-off-by: Angelica Ochoa <15623749+ao508@users.noreply.github.com>